Skipped performing registry issue checks on non-WinNT Operating Systems
A `System.NotSupportedException : This Operating System is not supported for issue checks to be performed` exception would be thrown on non-WinNT Operating Systems, but not caught. Instead of catching the exception, a check has been made to see if the registry issue checks are supported by the Operating System before trying to perform them.