Checked all issue files are collected, regardless of order
As long as all of the issue files are collected, it doesn't matter which order they are returned in. Commit fdc1925c605e169f94ee30bce6fcaa081e410381 added registry issue files to the list of issue files, but some Operating Systems may return the list of available JSON issue files in a different order to what is given in the test file.

`CollectionAssert.AreEquivalent` is used instead of `Assert.AreEqual` as `AreEquivalent` doesn't regard the order of the items.