Added ability to see if the registry value should match the issue file value
This is useful when seeing if something has been changed from the "current" value, without knowing what it is going to be changed to.

An example is seeing if the password to an account has been changed. It's not possible to know what the password is going to be set to, but if the registry value in the issue file doesn't match what is currently in the registry, the password has most likely been updated.